optional `:servicenet` config setting
The optional config setting `:servicenet` defaults to false.

When true, the setting will use the first private address of the
node instead of the first public address.  The first private ip
address should always be the ServiceNet interface (10.184.0.0/13
or 10.208.0.0/12).

The ServiceNet interface may be preferrable when running kitchen
on another Rackspace cloud server, like a Jenkins build server,
or for customers with VPN access to ServiceNet.

Using ServiceNet avoids bandwidth charges and may avoid changes
to the PublicNet/RackConnect firewall rules to allow
SSH connections to the kitchen test VMs from the public Internet.